python/recognition_handwritten_digital/pyproject.toml
[project]
name = "recognition-handwritten-digital"
version = "0.1.0"
description = "MNIST handwritten digit recognition"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "gradio",
    "numpy",
    "pillow",
]

[tool.uv]
